feat(render): preset "Bolígrafo grueso" — trazo cargado de tinta
La tinta extra sale del ÁREA del trazo (ink_width_jitter dilata el alpha), no
del gamma: ink_boost satura en +16% porque el núcleo del trazo ya es opaco y el
gamma sólo puede empujar los píxeles de borde del antialiasing.

- STYLE_PRESETS["Bolígrafo grueso"]: deriva de "Bolígrafo" (misma tinta azul-negra
  y misma mano) + ink_boost 0.20, ink_bleed 0.8, ink_width_jitter 0.35.
- ink_bleed baja de 1.8 a 0.8: el sangrado ya ensancha, y sumarle dilatación
  encima emborrona el trazo y cierra los ojos de las letras (verificado a 300 DPI).
- ink_bleed va en px ABSOLUTOS ⇒ la carga depende del DPI: ≈1.75× la masa de tinta
  de "Limpio" a 150 DPI y ≈1.53× a 300 DPI (seeds 42/7/101, masa umbralizada —
  el grano del papel es una constante aditiva que sesga el ratio hacia 1).
- writer_tab.py: el desplegable de estilos es una lista hardcodeada, no sale de
  STYLE_PRESETS; se añade ahí para que el preset aparezca en la app.
- 5 tests: contrato del preset, propagación por apply_style, y comportamiento
  (deposita más tinta que el default; apagar la dilatación se la quita).
- fix de lint arrastrado de R18: int(round()) redundante en renderer.py.

+++ core/inkcore/renderer.py
@@ -540,7 +540,7 @@ class HandwritingRenderer(BackgroundMixin, GlyphLoadMixin, LayoutMixin):
                     gi = page_start + k - 1
                     fat_line = self._fatigue_at(options, line_off=gi - self._doc_line)
                     if fat_line > 0.10:
-                        y_extra += int(round(fat_line * 9.0))
+                        y_extra += round(fat_line * 9.0)
                     x = options.margin_left_px + self._next_margin_offset(options)
                     if first_flags[page_start + k - 1]:
                         x += self._next_indent_px(options)
